backups/V3.3_Stable_MissedOrder/requirements.txt
MetaTrader5
requests
python-dotenv
redis
psycopg2-binary
